+++ pkgs/test_core/CHANGELOG.md
@@ -149,6 +149,8 @@
 * Populate `languageVersionComment` in the `Metadata` returned from
   `parseMetadata`.
 
+* Printing to stdout when using Runner is configurable
+
 ## 0.3.4
 
 * Fix error messages for incorrect string literals in test annotations.